% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amariSorting.R
\docType{data}
\name{amariSorting}
\alias{amariSorting}
\title{25 assessors twice sort and describe 12 amaris (i.e., bitter)}
\format{a list containing 3 objects:
 1) \code{Sorting} a data frame with the 12 by 25*2 sorting data,
2) \code{cubeOfVocabulary}: and array of dimensions
a 12 products * 41 words  (vocabulary) * 50 assessors-repetition
(i.e., 25 assessors *2 repetitions), and 3)
a data frame:
\code{information4Amaris} storing the description of the amaris.}
\source{
Lahne, J., Abdi, H., & Heymann, H. (2018). 
#' Rapid sensory profiles with DISTATIS 
and barycentric text projection: 
An example with amari, bitter herbal liqueurs. 
\emph{Food Quality and Preference}, 66, 36-43. 
(available from \url{www.utdallas.edu/~herve}).
}
\usage{
data("amariSorting")
}
\description{
\code{sortingAmari}: 25 assessors
twice sort and describe  12 amaris 
(i.e., bitter).
The data consist in
a list containing 3 objects:
 1) \code{Sorting} a data frame with the 12 by 25*2 sorting data,
2) \code{cubeOfVocabulary}: an array of dimensions
a 12 products * 41 words  (vocabulary) * 50 assessors-repetition
(i.e., 25 assessors *2 repetitions), and 3)
a data frame:
\code{information4Amaris} storing the description of the amaris.
}
\details{
The assessors are described by their 5 character
names with the following code: the first letter
{m/f} give the gender of the assessors (male versus female),
the second and third characters go from 01 to 25 and 
 uniquely identify the assessor, 
 the fourth and fifth characters
 identify the repetition (r1 vs. r2).

Some words of the vocabulary have been shortened for convenience;
here is the list of 
the short and long versions of the descriptors that have shortened:
'coffee_chocolate' <- 'coffee',
fortified wine' <- 'wine',
'red fruit' <- 'red',
'soy sauce' <- 'soy',
spirit (green)'  <- 'spirit',
'vegetal_green'  <- 'vegetal', and
'warm spice' <- 'spice'.

In the data sets, the amaris are identified with shortened
names, the whole names can be found in the data frame
\code{information4Amaris}.
}
\references{
Lahne, J., Abdi, H., & Heymann, H. (2018). 
Rapid sensory profiles with DISTATIS 
and barycentric text projection: 
An example with amari, bitter herbal liqueurs. 
\emph{Food Quality and Preference}, 66, 36-43.
}
\author{
Jacob Lahne, Herve Abdi, & Hildegarde Heymann
}
\keyword{DistatisR}
\keyword{datasets}
