% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeersProjectiveMapping.R
\docType{data}
\name{BeersProjectiveMapping}
\alias{BeersProjectiveMapping}
\alias{BeersProjectiveNapping}
\title{7 (fictitious) assessors evaluate 7 Beers using
\emph{Projective Mapping}.}
\format{a list with 2 elements:
1) \code{ProjectiveMapping}: a matrix of dimensions 7 beers by  7*2 
assessors-dimensions of the coordinates of the beers on the sheet
of paper; and
2) \code{Vocabulary} a list where each element of the list 
stores
the words used to describe a beer 
(words are separated with spaces).}
\source{
Abdi, H,  & Valentin, D. (2007). 
\url{www.utdallas.edu/~herve}
}
\usage{
data("BeersProjectiveMapping")
}
\description{
\code{BeersProjectiveMapping}:
7 (fictitious) assessors evaluate 7 Beers using
Projective Mapping.
}
\details{
First, Each assessor positionned the 7 beers on a sheet
of paper according the perceived similarity between the beers.
For each assessor, the position of the beers is recorded
from the \eqn{X} and \eqn{Y} coordinates. 
Second, the assessors are asked if they can describe the
beers with some freely chosen descriptors. These descriptors
are stored in a list with 7 elements (one per assessor)
where each element of the list
is a 7 component vector (one per beer) where each element stores
the words used to describe a beer 
(words are separated with spaces).
}
\references{
Abdi, H., & Valentin, D., (2007). 
Some new and easy ways to describe, compare, 
and evaluate products and assessors. 
In D., Valentin, D.Z. Nguyen, L. Pelletier (Eds) 
\emph{New trends in sensory evaluation 
of food and non-food products}.
Ho Chi Minh (Vietnam): 
Vietnam National University & Ho Chi Minh City Publishing House. 
pp. 5-18.
}
\author{
Herve Abdi
}
\keyword{DistatisR}
\keyword{datasets}
