% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP2SUMPCAnormedCP.R
\name{CP2SUMPCAnormedCP}
\alias{CP2SUMPCAnormedCP}
\title{CP2SUMPCAnormedCP}
\usage{
CP2SUMPCAnormedCP(CP3)
}
\arguments{
\item{CP3}{A 3D array of cross-product matrices}
}
\value{
The 3D array of the normalized cross-product matrices.
}
\description{
SUMPCA normalizes a cube of cross-product matrices.
}
\examples{
\dontrun{
D3 <- array(c(0, 1, 2, 1, 0, 1, 2, 1, 0,
              0, 3, 3, 3, 0, 3, 3, 3, 0), 
           dim = c(3, 3, 2))
CP2SUMPCAnormedCP(D3)         
}
}
