% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DblCenterDist.R
\name{DblCenterDist}
\alias{DblCenterDist}
\title{Double Center a distance matrix}
\usage{
DblCenterDist(Y)
}
\arguments{
\item{Y}{a "distance" matrix,}
}
\value{
a cross-product matrix (if the distance is Euclidean)
}
\description{
Double Center a distance matrix
}
\examples{
Y <- toeplitz(c(0, 3, 3))
DblCenterDist(Y)
}
