% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist2CP.R
\name{Dist2CP}
\alias{Dist2CP}
\title{Dist2CP}
\usage{
Dist2CP(D3)
}
\arguments{
\item{D3}{the cube of distance matrices.}
}
\value{
the cube of cross-product matrices.
}
\description{
Transforms a cube of distance matrices 
into a cube of cross-product matrices.
}
\examples{
\dontrun{
D3 <- array(c(0, 1, 2, 1, 0, 1, 2, 1, 0, 
              0, 3, 3, 3, 0, 3, 3, 3, 0), 
            dim = c(3, 3, 2))
Dist2CP(D3)
}
}
