% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistAlgo.R
\docType{data}
\name{DistAlgo}
\alias{DistAlgo}
\title{Four computer algorithms evaluate the similarity of six faces for distatis
analysis}
\format{
an \eqn{6\times6\times4}{6*6*4} array. Each \eqn{6\times 6}{6*6}
matrix is a distance matrix
}
\source{
Abdi et al. (2005). \url{www.utdallas.edu/~herve}
}
\description{
Provide the data.frame DistAlgo Data set to be used to illustrated the use
of the package \code{DistatisR}. Four algorithms evaluate the similarity
(i.e., distance) between six faces (3 females and 3 males). Each algorithm
provides a \eqn{6\times 6}{6*6} distance matrix evaluating the distance
between each pair of faces.
}
\references{
Abdi, H., Valentin, D., O'Toole, A.J., & Edelman, B. (2005).
DISTATIS: The analysis of multiple distance matrices.  \emph{Proceedings of
the IEEE Computer Society: International Conference on Computer Vision and
Pattern Recognition}. (San Diego, CA, USA). pp. 42--47.
}
\author{
Herve Abdi
}
\keyword{DistatisR}
\keyword{datasets}
