% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFAnormCP.R
\name{MFAnormCP}
\alias{MFAnormCP}
\title{MFAnormCP}
\usage{
MFAnormCP(Y)
}
\arguments{
\item{Y}{The matrix to normalize}
}
\value{
The normalized matrix
}
\description{
Normalizeq a positive semi-definite matrix matrix product such that its first eigenvalue is equal to one
}
\examples{
\dontrun{
A <- toeplitz(c(1, 0.6))
MFAnormCP(A)
}
}
