% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFAnormMat.R
\name{MFAnormMat}
\alias{MFAnormMat}
\title{MFAnormMat}
\usage{
MFAnormMat(Y)
}
\arguments{
\item{Y}{A matrix to be MFA-normalized}
}
\value{
An MFA-normalized Y
}
\description{
MFAnormMat
}
\examples{
MFAnormMat(Y)
}
