% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuclearnormCP.R
\name{NuclearNormedCP}
\alias{NuclearNormedCP}
\title{NuclearNormedCP}
\usage{
NuclearNormedCP(Y)
}
\arguments{
\item{Y}{The matrix to normalize}
}
\value{
The normalized matrix
}
\description{
Normalizes a positive 
semi-definite matrix by diving it by its
nuclear norm (i.e., the sum of the square root of
its eigen-values).
}
\examples{
\donttest{
A <- toeplitz(c(1, 0.6))
NuclearNormedCP(A)
}
}
