% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beersBlindSorting.R
\name{print.beersBlind}
\alias{print.beersBlind}
\title{Change the print function for class 'beersBlind'}
\usage{
\method{print}{beersBlind}(x, ...)
}
\arguments{
\item{x}{an object of the class 'beersBlind'}

\item{...}{the rest}
}
\description{
Change the print function for class 'beersBlind'
}
\author{
Herve Abdi
}
