% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Cmat.R
\name{print.Cmat}
\alias{print.Cmat}
\title{Print C matrix results}
\usage{
\method{print}{Cmat}(x, ...)
}
\arguments{
\item{x}{a list that contains items to make into the Cmat class.}

\item{...}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print C matrix results.
}
\author{
Derek Beaton
}
\keyword{internal}
\keyword{print}
