% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCubeOfCovDis.R
\name{print.cubeOfCovDis}
\alias{print.cubeOfCovDis}
\title{Change the print function for the class \code{cubeOfCovDis}}
\usage{
\method{print}{cubeOfCovDis}(x, ...)
}
\arguments{
\item{x}{a list objects of the class {cubeOfCovDis}}

\item{...}{the rest}
}
\description{
Change the print function for the class \code{cubeOfCovDis}: 
(output of \code{createCubeOfCovDis}.
}
\author{
Herve Abdi
}
\keyword{internal}
