% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beersBlindSorting.R
\name{print.cubeSort}
\alias{print.cubeSort}
\title{Change the print function for cubeSort}
\usage{
\method{print}{cubeSort}(x, ...)
}
\arguments{
\item{x}{an object of the class cubeSort}

\item{...}{the rest}
}
\description{
Change the print function for cubeSort
}
\author{
Herve Abdi
}
\keyword{internal}
