% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amariSorting.R
\name{print.dataAmari}
\alias{print.dataAmari}
\title{Change the print function for the data set: 
\code{dataAmari}}
\usage{
\method{print}{dataAmari}(x, ...)
}
\arguments{
\item{x}{a list: the data set from  {amariSorting}}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\value{
invisible; contents are printed to screen
}
\description{
Change the print function for the data set: 
\code{dataAmari}
}
\author{
Hervé Abdi
}
\keyword{internal}
