% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.df.excel.R
\name{print.distatis.data}
\alias{print.distatis.data}
\title{Change the print function for objects of 
class \code{distatis.data} output of\code{read.df.excel}}
\usage{
\method{print}{distatis.data}(x, ...)
}
\arguments{
\item{x}{a list: output of \code{distatis.data};
class \code{distatis.data}.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\value{
invisible; contents are printed to screen
}
\description{
Change the print function for bootRatios
}
\author{
Hervé Abdi
}
\keyword{internal}
