\name{print.DistatisR}
\alias{print.DistatisR}
\title{Print DistatisR results}
\description{
	Print DistatisR results.
}
\usage{
	\method{print}{DistatisR}(x,\dots)
}
\arguments{
  \item{x}{a list that contains items to make into the DistatisR class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton}
\keyword{print}
