% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.DistatisR.R
\name{print.DistatisR}
\alias{print.DistatisR}
\title{Print DistatisR results}
\usage{
\method{print}{DistatisR}(x, ...)
}
\arguments{
\item{x}{a list that contains items to make into the DistatisR class.}

\item{...}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print DistatisR results.
}
\author{
Derek Beaton
}
\keyword{internal}
\keyword{print}
