% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectVoc.R
\name{print.F4voc}
\alias{print.F4voc}
\title{Change the print function for objects of the class 
\code{'F4voc'} (e.g., output from function 
\code{projectVoc}).}
\usage{
\method{print}{F4voc}(x, ...)
}
\arguments{
\item{x}{a list: the data set: {str_BeersProjectiveMapping}}

\item{...}{the rest}
}
\description{
Change the print function for objects of the class 
\code{'F4voc'} (e.g., ooutput from function 
\code{projectVoc}).
}
\author{
Herve Abdi
}
\keyword{internal}
