% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePartial4Groups.R
\name{print.F_alpha_k}
\alias{print.F_alpha_k}
\title{Print F_alpha_k results}
\usage{
\method{print}{F_alpha_k}(x, ...)
}
\arguments{
\item{x}{a list that contains items for
the function \code{computePartial4Groups}
 output:
\code{F_alpha_k} class.}

\item{\dots}{inherited/passed arguments 
for S3 print method(s).}
}
\value{
invisible; contents are printed to screen
}
\description{
Print F_alpha_k results
}
\author{
Hervé Abdi
}
\keyword{internal}
\keyword{print}
