% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Splus.R
\name{print.Splus}
\alias{print.Splus}
\title{Print S+ matrix results}
\usage{
\method{print}{Splus}(x, ...)
}
\arguments{
\item{x}{a list that contains items to make into the Splus class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print S+ matrix results.
}
\author{
Derek Beaton
}
\keyword{print}
