% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeersProjectiveMapping.R
\name{print.str_BeersProjectiveMapping}
\alias{print.str_BeersProjectiveMapping}
\title{Change the print function for the data set: 
\code{str_BeersProjectiveMapping}}
\usage{
\method{print}{str_BeersProjectiveMapping}(x, ...)
}
\arguments{
\item{x}{a list: the data set: {str_BeersProjectiveMapping}}

\item{...}{the rest}
}
\description{
Change the print function for the data set: 
\code{str_BeersProjectiveMapping}
}
\author{
Herve Abdi
}
