% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary2CT.R
\name{print.voc4distatis}
\alias{print.voc4distatis}
\title{Change the print function for voc4distatis}
\usage{
\method{print}{voc4distatis}(x, ...)
}
\arguments{
\item{x}{a list: output of \code{vocabulary2CT}}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\value{
invisible; contents are printed to screen
}
\description{
Change the print function for voc4distatis class
 objects
 (e.g.,output of Rv).
}
\author{
Hervé Abdi
}
\keyword{internal}
