% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortingBeer.R
\docType{data}
\name{SortingBeer}
\alias{SortingBeer}
\alias{Sort}
\title{Ten Assessors sorted eight beers for \code{distatis} analysis}
\format{a data frame file containing 10 columns, 8 rows plus the names of
the rows and the columns.}
\source{
Abdi et al. (2007). \url{www.utdallas.edu/~herve}
}
\description{
Provide the data.frame \code{Sort}: Data set to be used to illustrated the
use of the package \code{DistatisR}. Ten assessors sorted eight beers.
These data come from the Abdi et al.' (2007) paper in \emph{Food Quality and
Preference}.  Each column represents the results of the sorting task for one
assessor.  Beers with the same number were sorted together.
}
\references{
Abdi, H., Valentin, D., Chollet, S., & Chrea, C. (2007).
Analyzing assessors and products in sorting tasks: DISTATIS, theory and
applications.  \emph{Food Quality and Preference}, \bold{18}, 627--640.
}
\keyword{DistatisR}
\keyword{datasets}
