% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortingSpice.R
\docType{data}
\name{SortingSpice}
\alias{SortingSpice}
\alias{SortSpice}
\title{21 French assessors sorted 16 blends of Spice for \code{distatis} analysis}
\format{a data frame file containing 21 columns, 16 rows plus the names of
the rows and the columns.}
\source{
Chollet et al. (2013). \url{www.utdallas.edu/~herve}
}
\description{
Provide the data.frame SortSpice: Data set to illustrate the use of the
package \code{DistatisR}. 
Assessors are sorting spices.
Each column represents the results of the sorting task for one assessor.
Spices  with the same number were sorted together.
}
\references{
Chollet, S., Valentin, D., & Abdi, H. (in press, 2013).  The
free sorting task. In. P.V. Tomasco & G. Ares (Eds), \emph{Novel Techniques
in Sensory Characterization and Consumer Profiling}.  Boca Raton: Taylor and
Francis.
}
\keyword{DistatisR}
\keyword{datasets}
