% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortingSpices.R
\docType{data}
\name{sortingSpices}
\alias{sortingSpices}
\title{62 assessors from 5 countries sort 16 spice samples}
\format{A list containing 7 objects
(for all sorting data, the number 
at the intersection of a row and a colum
indicates the number of the pile in which the spice was sorted):
 \code{}: A data frame containing the
1) \code{sortAll}: A data frame containing the
results of the sorting task for
all 62 participants,
2) \code{sortAmerican}: A data frame containing the
 results of the sorting task for
the 9 American participants,
3) \code{sortFrench}: A data frame containing the
 results of the sorting task for
the 21 French participants,
4) \code{sortIndian}: A data frame containing the
 results of the sorting task for
the 15 Indian participants,
5) \code{sortSpanish}: A data frame containing the
 results of the sorting task for
the 11 Spanish participants,
6) \code{sortVietnamese}: A data frame containing the
 results of the sorting task for
the 6 Vietnamese participants, and
7) \code{spicesDescription} A data frame containing the
description of the Spices.}
\usage{
data("sortingSpices")
}
\description{
\code{sortingSpices}: 62 participants from 5 different countries
(USA, France, India, Spain, and Vietnam) for 16 different spices
(including 6 mixtures of spices).

 The data consist in
a list containing 7 objects
(for all sorting data the number 
at the intersection of a row and a colum
indicates the number of the pile in which the spice was sorted);):
 \code{}: A data frame containing the
1) \code{sortAll}: A data frame containing the
results of the sorting task for
all 62 participants,
2) \code{sortAmerican}: A data frame containing the
 results of the sorting task for
the 9 American participants,
3) \code{sortFrench}: A data frame containing the
 results of the sorting task for
the 21 French participants,
4) \code{sortIndian}: A data frame containing the
 results of the sorting task for
the 15 Indian participants,
5) \code{sortSpanish}: A data frame containing the
 results of the sorting task for
the 11 Spanish participants,
6) \code{sortVietnamese}: A data frame containing the
 results of the sorting task for
the 6 Vietnamese participants, and
7) \code{spicesDescription} A data frame containing the
description of the Spices.
}
\details{
In the data frames, the spice blends are identifed
wiht acronyms that are expended in the data frame
\code{spicesDescription}.
}
\references{
Part of these data (i.e., the French sample) is described 
and analyzed in 
 Chollet, S., Valentin, D., & Abdi, H. (2014). 
 Free sorting task. In P.V. Tomasco & G. Ares (Eds), 
 \emph{Novel Techniques in Sensory Characterization 
 and Consumer Profiling}. 
 Boca Raton: Taylor and Francis. pp 207-227.
}
\author{
Chollet, S., Valentin, D., & Abdi, H.
}
\keyword{DistatisR}
\keyword{datasets}
