% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statis.R
\name{statis}
\alias{statis}
\title{statis}
\usage{
statis(
  LaGrandeTable,
  Norm = "MFA",
  center = TRUE,
  scale = "SS1",
  RV = TRUE,
  nfact2keep = 3,
  compact = FALSE
)
}
\arguments{
\item{LaGrandeTable}{A list of observations 
\eqn{\times}{*} variables matrices. These matrices describe the same
observations with sets of variables (which do not necessarily have 
to match across matrices).}

\item{Norm}{Type of normalization 
used for each matrix.  
Current options are \code{NONE} (do nothing), 
\code{SUMPCA} (normalize by the square root of the total sum of squares) 
or \code{MFA} (\code{default}) that normalizes each matrix so
that its first eigenvalue is equal to one.}

\item{center}{If \code{TRUE}, 
(\code{default}) the columns of each matrix are centered
(i.e., to have a mean of 0).
If \code{FALSE}, the columns will \emph{not} be centered.}

\item{scale}{If \code{TRUE}, the columns of each matrix are normalized
to have a standard deviation equal 1.
If \code{"SS1"}, (\code{default}) the columns of each matrix are 
normalized to have a sum of squares of 1, 
so that they contribute equally.
If \code{FALSE}, the columns will \emph{not} be normalized.}

\item{RV}{if \code{TRUE} (\code{default}) 
we use the \eqn{R_V}{Rv} coefficient to
compute the \eqn{\alpha}{weights}, 
if \code{FALSE}
we use the matrix scalar product}

\item{nfact2keep}{(default: \code{3}) Number of factors 
to keep for the computation of the
factor scores of the observations.}

\item{compact}{if \code{FALSE} (default),
 \code{statis} provides detailed output, if
 \code{TRUE},  \code{statis} sends back
only the \eqn{\alpha}{alpha} weights
 (this option is used to make the
bootstrap routine 
\code{\link{BootFromCompromise}} more efficient).}
}
\value{
\code{statis} sends back the results 
\emph{via} two lists:
\code{res.Cmat}
and \code{res.Splus}.
Note that items with a * are the only ones sent back
when using the \code{compact = TRUE} option.

\item{res.Cmat}{Results for the between 
          distance matrices analysis.}
\itemize{ 
  \item \code{res.Cmat$C}
  The \eqn{I\times I}{I*I} \bold{C} matrix
  of scalar products (or \eqn{R_V}{Rv} between matrices).  
  \item
  \code{res.Cmat$vectors} The eigenvectors of the \bold{C} matrix 
  \item
 \code{res.Cmat$alpha} * The \eqn{\alpha}{alpha} weights 
 \item
 \code{res.Cmat$value} The eigenvalues of the \bold{C} matrix 
 \item
 \code{res.Cmat$G} The factor scores for the \bold{C} matrix 
 \item 
  \code{res.Cmat$ctr} The contributions for \code{res.Cmat$G},
 \item
   \code{res.Cmat$cos2} The squared cosines for \code{res.Cmat$G}
 \item
 \code{res.Cmat$d2} The squared 
 Euclidean distance  for \code{res.Cmat$G}.
   }
   
\item{res.Splus}{Results for the between observation analysis.}
\itemize{
\item \code{res.Splus$SCP} an \eqn{I\times I\times K}{I*I*K} array.
Contains
the (normalized if needed)
cross product matrices corresponding to the
distance matrices.
\item \code{res.Splus$Splus} * The compromise 
(optimal linear
combination of the SCP's').
 \item \code{res.Splus$eigValues} *
  The eigenvalues of the compromise).
 \item \code{res.Splus$eigVectors} *
  The eigenvectors of the compromise).
\item \code{res.Splus$tau} * The percentage
of explained inertia of the eigenValues).
\item \code{res.Splus$ProjectionMatrix} The
projection matrix used to compute factor
scores and partial factor scores.
\item \code{res.Splus$Fi} The factor scores for the observations.
\item \code{res.Splus$Fj} The factor scores for the variables.
 \item 
  \code{res.Splus$ctr.i} The contributions for \code{res.Cmat$Fi}.
  \item 
  \code{res.Splus$ctr.j} The contributions for \code{res.Cmat$Fj}.
 \item
   \code{res.Splus$cos2.i} The squared cosines for \code{res.Cmat$Fi}.
 \item
   \code{res.Splus$cos2.j} The squared cosines for \code{res.Cmat$Fj}.
 \item
 \code{res.Splust$d2.i} The squared 
 Euclidean distance  for \code{res.Cmat$Fi}.
 \item
 \code{res.Splust$d2.j} The squared 
 Euclidean distance  for \code{res.Cmat$Fj}.
\item \code{res.Splus$PartialFi} an
\eqn{I \times \code{nf2keep} \times K}{I*nf2keep*K} array.
Contains the partial factors for the distance
matrices.}
}
\description{
multitable method with  "STATIS" optimization
procedure for asymmetric, rectangular matrices
}
\examples{
## Wine data from Abdi et al. (2012)
## A list of tables with matching rows and different variables
data(wines2012List)

## This replicate results from the paper             
statis.res <- statis(wines2012List, Norm = "SUMPCA", scale = TRUE, RV = FALSE)
}
