% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUMPCAnormCP.R
\name{SUMPCAnormCP}
\alias{SUMPCAnormCP}
\title{SUMPCAnormCP}
\usage{
SUMPCAnormCP(Y)
}
\arguments{
\item{Y}{Matrix to normalize}
}
\value{
Normalized matrix
}
\description{
Normalizes a positive semi-definite matrix (i.e., total intertia=1)
}
\examples{
\dontrun{
A <- toeplitz(c(1, 0.6))
SUMPCAnormCP(A)
}
}
