% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUMPCAnormMat.R
\name{SUMPCAnormMat}
\alias{SUMPCAnormMat}
\title{SUMPCAnormMat}
\usage{
SUMPCAnormMat(Y)
}
\arguments{
\item{Y}{A matrix to be SUM-PCA-normalized}
}
\value{
An SUM-PCA-normalized Y
}
\description{
SUMPCAnormMat
}
\examples{
SUMPCAnormMat(Y)
}
