% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementalProjection4distatis.R
\name{supplementalProjection4distatis}
\alias{supplementalProjection4distatis}
\title{Supplementary element projection in DISTATIS}
\usage{
supplementalProjection4distatis(res.distatis, elsupp)
}
\arguments{
\item{res.distatis}{the result of the function \code{distatis}}

\item{elsupp}{the supplementary elements (i.e., a 3D array).}
}
\value{
the coordinates of the supplementary 
and active elements in the RV space.
*** HA comment: Maybe we also want to send the square cosines 
to give the quality of representation.
}
\description{
\code{supplementalProjection4distatis}:
Computes for \code{distatis} the projection as supplementary element s
(a.k.a. "out of sample")
of a set of squared matrices. The matrices to be projected need to be
of the same type (i.e., distance, correlation) as the matrices 
used in with \code{distatis}.
}
\examples{
\dontrun{
# an Example here
}
}
\author{
Vincent Guillemot
}
