% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary2CT.R
\name{vocabulary2CT}
\alias{vocabulary2CT}
\title{Transforms a data.frame of product by vocabulary 
of assessors into
a  products by words (from vocabulary) contingency table.}
\usage{
vocabulary2CT(df.voc)
}
\arguments{
\item{df.voc}{a data frame with the vocabulary. In this 
data.frame 
each element 
stores
the words used by one assessor to describe a beer 
(words are separated with spaces);}
}
\value{
\code{CT.vocabulary} a matrix storing the 
prodcuts by words contingency table.
}
\description{
\code{vocabulary2CT}
 Transforms a data.frame of products by vocabulary 
of assessors into
a  products by words (from vocabulary) contingency table.
In  contingency table the number at the intersection
 of a row (product) and a column (word) is the number
 of assessors who used this word to describe that product.
}
\examples{
# Get BeersProjectiveMapping example
data("BeersProjectiveMapping")
aContingenyTable <- vocabulary2CT(BeersProjectiveMapping$Vocabulary)

}
\seealso{
\code{\link[tidytext]{unnest_tokens}}
 \code{\link[dplyr]{count}}
 \code{\link{vocabulary2CT}}
}
