\name{print.Cmat}
\alias{print.Cmat}
\title{Print C matrix results}
\description{
	Print C matrix results.
}
\usage{
	\method{print}{Cmat}(x,\dots)
}
\arguments{
  \item{x}{a list that contains items to make into the Cmat class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton}
\keyword{print}
