% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amariSorting.R
\name{print.dataAmari}
\alias{print.dataAmari}
\title{Change the print function for the data set: 
\code{dataAmari}}
\usage{
\method{print}{dataAmari}(x, ...)
}
\arguments{
\item{x}{a list: the data set from  {amariSorting}}

\item{...}{the rest}
}
\description{
Change the print function for the data set: 
\code{dataAmari}
}
\author{
Herve Abdi
}
\keyword{internal}
