% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiculturalSortingSpices.R
\name{print.dataSortingSpices}
\alias{print.dataSortingSpices}
\title{Change the print function for the data set: 
\code{multiculturalSortingSpices}}
\usage{
\method{print}{dataSortingSpices}(x, ...)
}
\arguments{
\item{x}{a list: the data set: {sortingSpices}}

\item{...}{the rest}
}
\description{
Change the print function for the data set: 
\code{multiculturalSortingSpices} 
(class \code{dataSortingSpices})
}
\author{
Herve Abdi
}
\keyword{internal}
