% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiculturalSortingSpices.R
\name{print.dataSortingSpices}
\alias{print.dataSortingSpices}
\title{Change the print function for the data set: 
\code{multiculturalSortingSpices}}
\usage{
\method{print}{dataSortingSpices}(x, ...)
}
\arguments{
\item{x}{a list: the data set: \code{sortingSpices}}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\value{
invisible; contents are printed to screen
}
\description{
Change the print function for the data set: 
\code{multiculturalSortingSpices} 
(class \code{dataSortingSpices})
}
\author{
Hervé Abdi
}
\keyword{internal}
