% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortingWines.R
\name{print.dataSortingWines}
\alias{print.dataSortingWines}
\title{Change the print function for the data set: 
\code{dataSortingWines}}
\usage{
\method{print}{dataSortingWines}(x, ...)
}
\arguments{
\item{x}{a list: the data set: {sortingWines}}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\value{
invisible; contents are printed to screen
}
\description{
Change the print function for the data set: 
\code{dataSortingWines}
}
\author{
Hervé Abdi
}
\keyword{internal}
