% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.df.excel.R
\name{print.distatis.data}
\alias{print.distatis.data}
\title{Change the print function for
class \code{distatis.data} outpu \code{read.df.excel}}
\usage{
\method{print}{distatis.data}(x, ...)
}
\arguments{
\item{x}{a list: output of \code{distatis.data};
class \code{distatis.data}.}

\item{...}{everything else for the functions}
}
\description{
Change the print function for bootRatios
}
\author{
Herve Abdi
}
