\name{print.Splus}
\alias{print.Splus}
\title{Print S+ matrix results}
\description{
	Print S+ matrix results.
}
\usage{
	\method{print}{Splus}(x,\dots)
}
\arguments{
  \item{x}{a list that contains items to make into the Splus class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton}
\keyword{print}
