% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary2CT.R
\name{print.voc4distatis}
\alias{print.voc4distatis}
\title{Change the print function for voc4distatis}
\usage{
\method{print}{voc4distatis}(x, ...)
}
\arguments{
\item{x}{a list: output of \code{vocabulary2CT}}

\item{...}{everything else for the function}
}
\description{
Change the print function for voc4distatis class
 objects
 (e.g.,output of Rv).
}
\author{
Herve Abdi
}
\keyword{internal}
