% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statis.R
\name{statis}
\alias{statis}
\title{statis}
\usage{
statis(
  LaGrandeTable,
  DESIGN,
  Norm = "MFA",
  center = TRUE,
  scale = "SS1",
  RV = TRUE,
  nfact2keep = 3,
  compact = FALSE
)
}
\arguments{
\item{Norm}{Type of normalization 
used for each cross-product matrix derived
from the distance (or covariance) matrices.  
Current options are \code{NONE}
(do nothing), \code{SUMPCA} (normalize by the total inertia) 
or \code{MFA} (\code{default}) that normalizes each matrix so
that its first eigenvalue is equal to one.}

\item{RV}{if \code{TRUE} (\code{default}) 
we use the \eqn{R_V}{Rv} coefficient to
compute the \eqn{\alpha}{weights}, 
if \code{FALSE}
we use the matrix scalar product}

\item{nfact2keep}{(default: \code{3}) Number of factors 
to keep for the computation of the
factor scores of the observations.}

\item{compact}{if \code{FALSE} (default),
 \code{statis} provides detailed output, if
 \code{TRUE},  \code{statis} sends back
only the \eqn{\alpha}{alpha} weights
 (this option is used to make the
bootstrap routine 
\code{\link{BootFromCompromise}} more efficient).}

\item{LeListOfTables}{an "observations 
\eqn{\times}{*} observations
\eqn{\times}{*} distance matrices" array of dimensions
\eqn{I\times I \times  K}{I*I*K}.
Each of the \eqn{K} "slices" is a \eqn{I\times I}{I*I} square
distance (or covariance) matrix describing the 
\eqn{I} observations.}

\item{Distance}{if \code{TRUE} (\code{default}) 
the matrices are distance matrices, \code{FALSE}
the matrices are treated as positive semi-definite matrices
(e.g., scalar products,
covariance, or correlation matrices).}

\item{double_centering}{if \code{TRUE} 
(\code{default}) the matrices are double-centered
(should always be used for distances).
if \code{FALSE} the matrices
will \emph{not} be double centered 
(note that these matrices 
should be semi positive definite matrices such that
covariance matrices).}
}
\value{
\code{statis} sends back the results 
\emph{via} two lists:
\code{res.Cmat}
and \code{res.Splus}.
Note that items with a * are the only ones sent back
when using the \code{compact = TRUE} option.

\item{res.Cmat}{Results for the between 
          distance matrices analysis.}
\itemize{ 
  \item \code{res.Cmat$C}
  The \eqn{I\times I}{I*I} \bold{C} matrix
  of scalar products (or \eqn{R_V}{Rv} between matrices).  
  \item
  \code{res.Cmat$vectors} The eigenvectors of the \bold{C} matrix 
  \item
 \code{res.Cmat$alpha} * The \eqn{\alpha}{alpha} weights 
 \item
 \code{res.Cmat$value} The eigenvalues of the \bold{C} matrix 
 \item
 \code{res.Cmat$G} The factor scores for the \bold{C} matrix 
 \item 
  \code{res.Cmat$ctr} The contributions for \code{res.Cmat$G},
 \item
   \code{res.Cmat$cos2} The squared cosines for \code{res.Cmat$G}
 \item
 \code{res.Cmat$d2} The squared 
 Euclidean distance  for \code{res.Cmat$G}.
   }
   
\item{res.Splus}{Results for the between observation analysis.}
\itemize{
\item \code{res.Splus$SCP} an \eqn{I\times I\times K}{I*I*K} array.
Contains
the (normalized if needed)
cross product matrices corresponding to the
distance matrices.
\item \code{res.Splus$Splus} * The compromise 
(optimal linear
combination of the SCP's').
 \item \code{res.Splus$eigValues} *
  The eigenvalues of the compromise).
 \item \code{res.Splus$eigVectors} *
  The eigenvectors of the compromise).
\item \code{res.Splus$tau} * The percentage
of explained inertia of the eigenValues).
\item \code{res.Splus$ProjectionMatrix} The
projection matrix used to compute factor
scores and partial factor scores.
\item \code{res.Splus$Fi} The factor scores for the observations.
\item \code{res.Splus$Fj} The factor scores for the variables.
 \item 
  \code{res.Splus$ctr.i} The contributions for \code{res.Cmat$Fi}.
  \item 
  \code{res.Splus$ctr.j} The contributions for \code{res.Cmat$Fj}.
 \item
   \code{res.Splus$cos2.i} The squared cosines for \code{res.Cmat$Fi}.
 \item
   \code{res.Splus$cos2.j} The squared cosines for \code{res.Cmat$Fj}.
 \item
 \code{res.Splust$d2.i} The squared 
 Euclidean distance  for \code{res.Cmat$Fi}.
 \item
 \code{res.Splust$d2.j} The squared 
 Euclidean distance  for \code{res.Cmat$Fj}.
\item \code{res.Splus$PartialFi} an
\eqn{I \times \code{nf2keep} \times K}{I*nf2keep*K} array.
Contains the partial factors for the distance
matrices.}
}
\description{
multitable method with  "STATIS" optimization
procedure for asymmetric, rectangular matrices
}
\examples{
add(1, 1)
add(10, 1)
}
